backend/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
tensorflow
torch
joblib
psycopg2-binary
pydantic
requests
